+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 See [changes since release][HEAD].
 
+* Add "aria-live" to tablists for reporting tab changes to the user.
+
 ## [3.2.0][] / 2018-03-07
 
 * No longer trigger selection actions during Component destruction.